Gets rid of monad writer
Given that errors are the name of the game, it seemed like there
was too much indirection to be using a writer, especially because it
required frequently manipulating the stored value in the monad.

By getting rid of this and using the stored value as the subject of the
monad, we are able to manipulate it more freely. This happens, though,
at the expense of having the monad hold anything other than the error
state.

The place where we see this is selectionToFields, which used to return
a list of fields _or_ an error if a field can't be found. Now, it just
returns a list of fields and pushes the error to the lookup against the
list.